Move EnhancedSwitch under internal
